Allow to manually specify the backend for an analysis (#5)
* Allow backend to be specified manually

This allows to use multiple backends side-by-side. If the backend is not specified, the analysis decorator will pass the default backend to the analysis.

* Modify tests for the analysis decorator

The order of the 'return_raw_result' and 'backend' parameters are now not checked anymore, it is sufficient to make them keyword-only parameters (which is enforced).