Improve GPU/Torch support for the Optiland backend (#75)
* Add GPU support to tests

* Convert ray tracing results to Numpy arrays before creating DataFrame

* Use use_gpu fixture in optiland_backend fixture

* Set torch default device when changing optiland backend device

* Default to float64 precision for torch backend

* Make Optiland's computation backend configurable via the --optiland-backend option

* Add tests for Torch backend to CI workflow

* Skip Huygens PSF tests for Torch backend

* Explicitly calculate indices for wavefront OPD map generation

This makes sure that the OPD map analysis works with all computation backends, regardless of the order in which pupil coordinates are generated.

* More explicit typing of visisipy.set_backend()

Instead of using the base BackendSettings class to specify settings for all backends, overloads are created for each backend type

* Make psf extent calculation compatible with torch backend

* Fix test error in test_build_retina_not_image_raises_valueerror

Optiland no longer allows to insert surfaces at index 0 if the system already contains a surface at index 0.

* Add torch to development dependencies

* Document how to initialize the Optiland backend with GPU acceleration

* Small documentation fixes

Fix an incorrect link and reference to the Optiland GPU backend documentation from the installation page

* Ensure scalar tensors are converted to floats

* Require Optiland version 0.5.8 or higher

* Use noninteractive backend for matplotlib in tests